Zig build for cross-compilation.
Includes compiling the zig code, assembling the .s files, and linking
using the custom linker script.

It even makes the binary kernel8.img file we need.

The only thing I can't figure out is how to get it to copy the output
from under the zig-cache directories to the project root.