Coding on the road.
Adding support for Github codespaces. This builds the container
with QEMU and a particular point-in-time version of Zig (dev
version of 0.11, to be specific.)

Someday we'll need to update this to keep up with Zig's rapidly
evolving API...

But for now, starting this codespace then running "make clean all"
is all it takes!