Debugging the real hardware.
Mailbox is returning different statuses.

There may be an alignment or a cache invalidation issue.

As a shot in the dark, I will try invalidating the data cache for the
memory occupied by the mailbox message. That means I need some more
assembly.

Preparing for that, I'm making the existing assembly more consistent
and sane.