Ziggy Forth is kind of working.
Something odd about the stack.

I need a clear screen routine for the frame buffer. Either that or the
ability to scroll when reaching the bottom of the screen.

'42 emit' panics due to i32 vs u32 mismatch.

Would be nice to recover from a Zig panic back to a prompt.