Really crude panic recovery.
Use the word `panic` to trigger a panic. We can come back to the
interpreter's top level prompt.

The stack grows by 5 frames each panic. Fixing that will require a
more realistic model of a process and process-switching. (Probably
also a good time to isolate the kernel stack from the application
stack, even if we're running everything at EL1.)