Fix bug triggered when returning an error
kernelInit() claims to be an error-returning function, but it is
usually "called" by a direct jump (eret to the address). So it does
_not_ have a proper Zig error return trace object in its stack. Making
it be a non-failable function causes Zig to generate the code to set
up that object which is then used (miles away) when a failable
function returns an error.