Start making hard-coded devices more robust
Beginning with the mailbox. We've been using hard-coded
addresses. That's not forward-compatible as these things move from one
RPi to the next.

The devicetree tells us where to find the registers. A node like
`/soc/mailbox@7e00b880` will have a `reg` property with the values
`0x7e00b880 0x40` to indicate there are 64 bytes of registers starting
at address 0x7e00b880 _on the physical bus_.

We need our device drivers to understand address translation from the
CPU to the bus and back. Previously this was _also_ hard-coded in the
bsp's `memory.zig` module. Now we will get it from the devicetree as
well. The `/soc` node has `ranges` that tell us how to translate from
the parent (CPU) address space to the child (bus) address space.