Migrating another device to the new model.
I've called this one "peripheral clocks"... it's not the same as a
timer or even a counter. These just control the hardware's own clock
rates. Not sure what the "generic" interface for this device should be
so I've added it only to the `bsp.raspi3` module without a
corresponding generic on `bsp`.