Nearing the end of implementing the new device model.
`bsp/raspi3` just has a couple of constants now, based on the device's
fixed memory locations. Perhaps these can be gotten from the
devicetree, but there's a chicken-and-egg problem: we need the memory
addresses to create the allocator, but we need the allocator to parse
the device tree.

GPIO is now relocatable and instantiated at init time. It is injected
into the Pl011 controller instead of having a hardcoded location.