Restruture the kernel entry stack frame
Used to have the stored FP and LR at the bottom of the stack
frame. Now it's at the top of the stack frame with the local structure
below it. This is more similar to an ordinary function invocation that
Zig would generate.

It didn't seem to help the bug, but it helps the debugging.