[Breaking] Forty can now change the foreground / background colors
`fb` used to return the address of the frame buffer memory. It now
returns the base of the FrameBuffer struct. All members of the
FrameBuffer struct are now visible to forty.

Since `buffer_size` is part of the FrameBuffer struct, I've removed
the `fbsize` word.

I've added fg and bg as members of the struct. They can be set like:

`5 fb fb.bg + !b`

Using `!` instead of `!b` will almost always result in an alignment
panic. (This might not always happen if we change the struct
definition.)