First cut of compile time configuration
New option to `zig build`: -Dboard

Enum `SupportedBoard` in build.zig defines the allowed
values. Function `configModule` in build.zig picks the source file.

The chosen source file (found under config/) will be made available to
the build as `@import("config")`. Notice that module doesn't have
".zig" at the end.

We should keep the logic & code in the `config/` files to a minimum,
it should probably just be simple values. We can then use `switch` or
`if` statements in the main codebase to take effect.