WIP commit: status before holiday / vacay / work trips
Much documentation added, in code within `dwc_otg_usb.zig` (in the
form of register definitions) and under `src/usb` (in the form of UML
diagrams and notes in markdown).

The driver doesn't exactly work... the first transfer results in a
'transfer halted' interrupt instead of completion.

The state machine is horribly complicated. It conflates Transfer and
Transaction and makes inferences based on booleans rather than
explicit states. Rewriting it is why I took all the notes under
`src/usb`.