Basic graphics primitives should not return errors.
Also, this restores the original intent of line-demo-handler. (It got
borked when I reversed the order of arguments to line and fill.)

As an interesting side effect, we can now measure the difference in
speed between DMA fill and CPU fill by commenting out the DMA fast
path in `fill`.