Switch back to an opcode based inner interpreter from the existing purely threaded model.  Also simplified the "word" function interface dropping the body and offset parameters. We can do this because the opcodes now take care of program flow. For the same reasons, word functions no long return an offset.
At the price of making the inner function more complicated
these changes declutters the many word functions and
eliminates a huge number of function calls. It also makes
it possible to generate a forth word stack trace.

Add the stacktrace word, tho it is currently buggy in that
it currently omits words executed by `eval` and `exec`.

Also renamed the rstack -> istack (Itermediate) since it
is no longer involved in returning anything.