Make 'const' not 'var'... part 1 of N.
Newer Zig versions have made it an error to declare something a var if
it is never mutated. These are easy changes, but we've got a bunch of
them.

Even though we've pinned ourselves to a particular 0.12 nightly (see
README) it makes sense to clean these up so any eventual update isn't
a nightmare.