Improved debugging for USB initialization
Some debug logs moved from mring to serial console. They interleave
with Qemu traces (sometimes interrupting each other) so we can see
what the virtual HW thinks is happening at different stages.

Also, the forth word 'usb-init' is replaced by two words:

`usb-init-hcd` should be run just once to set up the chip itself. This
actually works.

`usb-init-root-port` creates a Device for the root port on the
chip. This is the first place we need to do actual transfers (the "get
descriptor" control transfer to endpoint 0.) This is also where things
stop working. It _looks_ like the first two stages of the transfer
complete but then we get a stall instead of a status packet at the
last stage.