First successful interrogation of a device!
`usb-init-root-port` is now able to successfully get the device
descriptor from Device 0, Endpoint 0. (The root hub embedded in the
SoC.)

Unfortunately the device descriptor is not easy to get from Forty
right now... you have to read the Qemu traces to see the address shown
in the HCDMA register and dump it. Better interop coming soon.