USB Hub can now report on itself.
New word `usb-init-bus` returns the address of a `usb.Bus` struct. At
the moment, that struct's only member is a `usb.Hub` which is the root
of the device tree.

On initializing the bus, the hub is also initialized which includes
getting the HubDescriptor. If debug is enabled for the .usb log scope,
then the HubDescriptor will be dumped on the console.