Another massive commit... marathon hacking session
*Still doesn't work on HW, WIP on the root_hub implementation*

- New module TransferFactory constructs all kinds of control
transfers (in the future will be extended to interrupt and bulk
transfers as well)

- All the flavors of setupXXXPacket are consolidated in
TransferFactory.

- A device's full configuration can be parsed now. One use can be seen
in the Device struct at the bottom of dwc_otg_usb.zig. The parser is
in usb/device.zig. Much fighting with Zig's struct layout and ARM's
alignment constraints in this code.

- The HCI has learned a new function `perform` which performs an
entire USB Transfer. (Only supports Control transfers for now.) This
handles all the stages, states, and results of the individual
transactions. It dispatches transactions to channels according to the
state of the Transfer.

- There is still some mixing of USB protocol specified behavior and
DWC-specific behavior in `perform`. This might be factored out if I
can figure out how to do it without polymorphism.

- Transfer has it's own state machine. Whenever a transaction
completes (or aborts, or errors, etc.) the Transfer is informed and
updates its state accordingly.

- It is likely that there will also be a `perform` function on
Endpoint, which will attach the endpoint details and device address to
the transfer before handing it off to the HCI. Until then, we must
remember to set the transfer's `device_address` every time.