Intermediate commit with new Interop mechanism
- Any function with an argument of type forty.auto.InteropCall can be
exposed to forth.

- forty.auto.defineNamespace is the entry point to do that.

- examples can be seen in frame_buffer.zig with the defineModule
function and in usb.zig with its defineModule function.

- Integer, pointer, and string arguments are converted automatically
to and from forth u64 values. Enums might work, if you're calling and
returning the same kind of enum.

- Functions can return errors or nulls.

- Any function that returns an error will end up with 2 values on the
stack. TOS will be the actual return value... it will be 0 if there
was an error. Next on the stack is the error value... _it_ will be 0
if there was not an error.

- The exception is a function that returns `!void`. That will have
only the error value as TOS, which will be 0 to indicate no error and
any non-zero value means there was an error. (The number is actually
Zig's error enum value, but that can change with every compilation so
it's best not to depend on it.)

- Signed integers are OK to return. They get sign-extended and
bitcasted into a u64's bit pattern. I have no idea how to get Zig to
let me coerce them back from u64 into the right smaller bitpattern to
pass as a parameter.