Updating the USB init sequence.
- `usb.init` :: Initialize the whole USB subsystem, from the main bus
topology to the HCI
- `usbhci` :: Pushes the address of the HCI module (currently it will
always be the dwc_otg_usb driver)
- `usbhci.initialize` :: Initialize the DWC chip (host controller
interface) only.
- `usbhci.rootPortInitialize` :: Initialize the root port ... WARNING
this is the part that doesn't work and will soon be replaced with the
fake root hub.

`usbhci.initialize` is equivalent to the old `usb-init-hcd` (but with
less VTable). `usbhci.rootPortInitialize` replaces
`usb-init-root-port`. There is no replacement for `usb-init-bus` which
never worked.

The HCI functions are not as easy to type, and I've been typing them a
lot. `usb.init` should take the place of both of those.