Improve backtraces inside exception handlers
GDB doesn't fully respect the Aarch64 Procedure Call
Standard (PCS). It should use the Frame Pointer register (x29) to find
the current function's activation frame, then use that as a linked
list to previous frames.

Instead it assumed the activation frame is always at the bottom of the
stack. I guess that's how GCC builds its frames so that's how GDB
expects to find them.

Anyway, this comment reorganizes the stack on entry to an
exception. The activation frame is now at the bottom of the stack with
the "context structure" of saved registers just above that.