[bugfix] 77 was being left on the stack after boot
In the interop changes, I accidentally overwrote the definition of
`clear`. So the tests in `init.f` were attempting to clear the stack
but were actually trying to clear the screen (repeatedly).

Except that they (of course) weren't passing the framebuffer pointer
to the interop version of `clear`. I misunderstood the intent of clear
when making the interop changes and added _another_ redefinition of
`clear` at the top of init.f as an alias for `cls`.

As a result of `clear` _not_ clearing the stack, `test-if` left the
value 77 in TOS after the first test stanza.