!void should only push its error value
Previously, a !void function would leave a zero on the stack on
success, but on error would leave `e 0x7fff_ffff_ffff_ffff` where `e`
is the int value of the error. (An `e` of 0 means "no error".)

The intended contract for error union return types is that they
leave the same number of values on the stack whether the return is an
error or a good value. For non-void returns that means two values on
the stack `e v`. For void returns, it means `e`.