Refactoring auto struct and namespace definition
There's no need for callers to reach into the `auto` module. We can
call it directly on the Forth instance.

Also, I found a way to unify defineStruct and defineStructRecursive,
by using a comptime struct of options (this is a somewhat common
pattern in the Zig stdlib.) The options specify whether to expose
decls from the module and whether to recurse into contained
structs. Some of our modules don't work with recursion on for reasons
I don't understand, so it's an opt-in feature that is only enabled in
a couple of places where I know it works.