Some cleaning of root
I finally figured out how to make Zig happy with panic being in
debug.zig instead of main.zig.

Zig's stdlib looks for a decl with the name 'root.panic', so it's
sufficient to have a pub const in root, even if the _value_ of it is
just an alias to debug.panic.