WIP commit: thread sleep
Partially implemented. Reveals a problem with threads as they exist
today... we can't start the scheduler until after there's at least one
thread, but once there _is_ one thread, reschedule will hop over to
it. This means that in the (currently not-a-thread) kernelInit, we
have to create & ready the first thread before starting the timer, but
once we create & ready the first thread, we're not in kernelInit any
more to start the timer.

The likely solution is to transmute kernelInit into a thread by
directly manipulating the thread table entries.