Thread sleep implemented.
Required some changes to kernel operation. Now the kernel's main line
of execution is registered as thread 0. Whenever there's nothing else
to do--meaning all other threads are sleeping or waiting for
IO--thread 0 will get scheduled. It's always in the ready state, but
it does nothing.

There's also a new high-level interface for spawning a thread. It's
more Zig-friendly than the low level one that requires a bunch of
@intFromPtr gunk.