Shore up interrupt masking.
It's not a good idea to have two different ways to mask and restore
interrupt flags. Mixing the synchronize.criticalSection and
the (simpler) cpu.disable mechanisms was causing hangs and interrupt
retriggering.

This commit removes the criticalSection mechanism and uses the
cpu.disable one instead.

Both mechanisms are flawed when we run multiple cores, however. They
only prevent the same core from re-entering a block. We will need a
better mutex soon.