Scheduling now appears stable, with one thread at least
I narrowed down the problem that was causing the hang. Apparently,
testing the handler value against `undefined` is not reliable. The
clock handler was not getting called, which meant two things: the
timer interrupt bit was never cleared so we'd frequently hang at boot
in an IRQ loop; and second, if we somehow got through boot, the clock
counter would not get reset so we'd never see another scheduling
increment.

Apparently, the solution here was to use `null` as the value to
indicate "no handler" instead of `undefined`. I narrowed down every
other change to just that difference. (There are some cosmetic and
defensive programming changes in this commit as well, but the change
from `undefined` to `null` and the change from `*IrqHandler` to
`?*IrqHandler` is the key.)