Squash commit of the wip-usb-with-threads branch
Changes include:

- Enums vs Constants: Moved from enums to constants for fields in
protocols. New rule of thumb for style:

- When interacting with a protocol (device, memory-mapped register, or
interop struct), use a struct with nested constants to represent the
allowed values for a field.

- Use enums only for internal (i.e., logical) applications.

- Consider an enum with assigned values in the members as a yellow
flag... it may indicate a protocol interaction downstream.

- Use of @intFromEnum and @enumFromInt are also yellow flags

- Updated some instances where `var` was used unnecessarily, changing
them to `const` in compliance with Zig 0.12's stricter checking

- Import styles in main.zig more consistent. Previously, we mixed
direct imports with root.* indirection.

- Improve readability and consistency of main.zig

- Fake root hub can detect and attach the QEMU hub.

- Introduced "Mailbox", a new synchronization construct implemented as
a generic Zig type. It is a blocking FIFO queue managed by a pair of
semaphores