First successful communication to 3rd layer device
host software -> root hub -> qemu hub -> qemu keyboard

Able to retrieve configuration descriptor, but there are some things
stopping us from going further.

1. The keyboard reports class-subclass-protocol as 0-0-0, which means
"read the interface descriptor" but we don't look into that yet.
2. The interface descriptor has interface class-subclass-protocol as
3-1-1, which means "HID, boot interface, keyboard"!
3. We don't have a driver that knows how to bind a keyboard, so the
init fails and we try to disable the port on the qemu hub. (And for
some reason that disable fails.)

But still... it means we got a "real" (emulated) hub talking to our
host software.