Scheduler should not allow starvation
We have threads that spin (like the Forty main thread, waiting for a
keystroke). If these are higher priority than background threads they
could starve the background. On the other hand, some of those
background threads are CPU hogs too.

The scheduler now debits the priority of a thread that exhausts its
quantum. Once the current priority is lower than that of a waiting
thread, the waiting thread will get scheduled.

When a thread wakes up from sleep or receiving a message, it's current
priority is restored to its base priority.