Idiom for "don't care" value of `*anyopaque` parameter broke
When we don't care what pointer is to be passed, but we can't use a
null pointer, we had been using `&.{}`. That now results in an error:

```
type 'anyopaque' does not support array initialization syntax
```

In some places, we can use `undefined` as the "don't care" value. For
calls to the `schedule.spawn` API function, I defined an official
"don't care" value called `schedule.no_args` that callers can pass
when they don't have any arguments to provide.