More of the new data model
No transfers have been harmed in the making of this commit.

There's a new pathway under construction, using Hub -> HubPort but it
is not used for any actual transfers yet. The next step is likely to
be disruptive, as I change the old style interrupt handling to a new
style that will (hopefully) be much clearer.