New data model complete.
USB is now working about as well as it did three weeks ago. The
difference is that this model is less device centric and more focused
on the endpoints and interfaces.

It also has a clearer tree structure for the bus.
Hub -> HubPort -> Hub.

Still not working on actual hardware, but with a brand new problem to
work out.