Keycodes now used for input
Keycode is a u16 that holds an ascii character in the lower 8 and some
modifier bits in the upper 8. See `key.zig` for keycode definitions.

For USB, `hid_keyboard.zig` translates usage codes directly into keycodes.

Serial (UART) input is translated according to a subset of ANSI escape
sequences defined in `term.zig`