BREAKING: `create` changed to parse its own word
This brings it into alignment with the forth standard. It also helps
prepare for implementing `does>`.

New word: `header` ( addr len -- )

Behaves like `create` used to. Takes the addr and len of a name on the
stack and adds a dictionary header for a new word. Updates LATEST and
HERE.

A zero length name is totally fine... it creates an anonymous
word. (See `:noname`)

`create` is now behaves as if it was defined like `: create word
header ;`