Update `see` to work with child words
On a normal high-level word, `see` is almost good enough to round-trip
a word. (Branch targets are the exception.)

For a word created by `variable`, there's now a special format from
`see` that shows the DFA and the current value of the variable.

For a child word created by a defining word that uses `does>`, `see`
now shows the DFA that will be pushed on the stack, a comment-like
sequence that shows the address of the "does> target" and the thread
of the parent word that provides the execution semantics of the
child. (Whew, what a mouthful.)

Example:

```
> see constant
: constant create , does> @ ;
> 17 constant foo
> see foo
: foo 0xb15e0 ( does a8b90 ) @ ;
```