Introducing the BIOS design.
* `include/asm/macros.h` has the new macro BIOS(n) to invoke a routine
from a BIOS vector table
* `src/arch/aarch64/bios.S` has routines to initialize the BIOS
table (I can't find a way to have it assembled absolutely and be at a
low address. The kernel is loaded from a binary image that is both
loaded at and starts execution at 0x80000.)
* `bios.S` now holds the UART init, in, and out functions. In and out
are the first two entries in the BIOS table.
* `video.f` currently has it's own versions of `cr`, `cls`, `emit`,
and `tell`. These should be temporary. Once we get the ability to
re-vector BIOS routines, they should go away.