The final purge of Zig
Remove 'zig cc' as build tool. We each need to install a
cross-compiler toolchain.

Move all old zig files to the attic. Some of them have useful hardware
knowledge embedded... we'll see if we use any of these files in the
next month or two. If not, we can delete them for good. (Git will have
them in case we _really_ need the info.)