First working version of find-file and cd
Some ugliness with taking string input, but we can locate the first cluster of a file now.

```
s" SRC        " find-file
```

The extra spaces are because I'm just directly matching the input string with the 8+3 name in the
directory entry. It'll get better later.

For `cd`, I'm trying something out. Not totally happy with it. I was trying to get the ability to do
`cd src` by having cd act like a parsing word. With the way parse is implemented now, plus the way
find-file matches filenames, cd has to look a little weird:

```
cd SRC        "
```

The trailing `"` after the spaces is not a typo.

However, given all that, `cd` will set the current directory or tell you the file wasn't found.