Add `defer` and `is` for vectored execution
- `defer` creates a word with no behavior
- `is` assigns an execution token to a word

Example:
```
defer greet
: say-hello ." Hello, world! " ;
: t1 ['] say-hello ;
t1 is greet
```

The need for the extra compiled word `t1` in that example results from
the "cheat" described in the definition of `'` inside armforth.S. The
current implementation of `'` means it only works when compiling a
definition. We might be able to remove that limitation in the future.