Attempt to make vectored execution easier.
TICK never behaved quite like standard forth. In particular, it would
not work in interpreted mode, only in a compiled word. This means that
nice constructs for making jump tables didn't work. For example:

```
create buttons
' tune-radio ,
' oil-slick ,
' missiles ,
' ejection-seat ,
```

would require a separate array allocation, a compiled initializer, and
an invocation of the initializer. Now TICK `'` behaves more like ANS
Forth so it puts the XT of the next word in the input stream onto the
stack. COMPILE-TICK `[']` acts like Jones' original `'`... it works in
a compiled word to put the XT of the next word in the definition onto
the stack.

This is illustrated with the toy application `whichway` that is now
included for use with `evaluate`.