Use a word with all 1 bits as the "true" value. "False" remains 0
[BREAKING]

We were using a C-like or Clojure-ish version of
truthy... 0 was false, anything else meant true.

There are a lot of useful Forth snippets in the wild, and some of them
assume the standard definition where a word full of 1 bits means
"true". Some of these snippets do things like bitwise-and an exception
number with a boolean value. Like this:

```
word find dup 0= -13 and throw
```

This uses several tricks at once. 0= yields a boolean, and-ing it with
-13 results in either 0 or -13. Throw only does anything if TOS is
non-zero, so this is a conditional abort. But it doesn't work if the
result of 0= is some random bit pattern that just happens to not be
zero.