Further departing from didactic path.
We have a few useful words that we'd like to use during boot:
defer, is, and postpone. However, these make use of if and throw. So
move if, then, else, throw, and catch into armforth.S as
defcode (secondary) words.

I've also updated a few definitions to be more "contemporary" (for
whatever that means in Forth at least!) For example, we can define
paren-comments using a constant for ')' and parse. For that matter,
since constant, value, and variable are all just applications of
does>, we can use them earlier on in the boot process too.

These changes simplify some of the earlier definitions in boot0 at the
cost of requiring the reader to know more forth earlier on.

The intermediate target is to use postpone for some of the control
structure definitions. (Especially the counted loops!)