Better (but not great) output when a test has incorrect result
Previously, the test's "incorrect result" line would show screens full
of garbage, due to `source` returning the a mismatched addr/len
pair. (Would return the keyboard buffer's addr but the length of the
string being `evaluate`d.) With the source addr fixed, the "incorrect
result" would print the _entire_ string being evaluated.

It's not obvious how to get just the last line of input when
evaluating. So, this is a little workaround (hack) to show up to the
last 60 chars of input. The failing case will be at the _end_ of the
output.