Interim checkpoint on the `here` redefinition
On the path to making `here` simply return the here address. (Today it
returns the address of the cell that holds the
address. Pointer-to-pointer.)

We modify here fairly often, so this introduces a temporary word
`&here` which is a pointer to here. In this first step, I haven't
redefined `here` yet. Just making sure we can use the indirection correctly.