Fix: video init on hardware was broken for a long time
I missed an address translation when converting the video init from
forth to asm. Qemu does _not_ accurately emulate the communication
from the ARM core to the VideoCore GPU. On real hardware, the GPU
reports addresses in it's own view of memory, which doesn't match the
ARM core's view.

On a Pi3, the difference is between $FEB3A000 and $3EB3A000, but
that's all it takes.